WIP working on restructuring (not compiling)
